requirements.txt
python-dotenv
twscrape
transformers
jax 
jaxlib 
flax
tf-keras
tensorflow-cpu
torch